moved contact form to profile_detail.html and deleted sendemail app - the form is missing for some reason - this requires debugging
