+++ public_profiles/forms.py
@@ -1,4 +1,6 @@
 from django import forms
+from django.contrib.auth.models import User
+from private_profiles.models import Profile
 
 class ContactForm(forms.Form):
     from_email = forms.EmailField(required=True)
