+++ public_profiles/templates/public_profiles/profile_details.html
@@ -1,0 +1,31 @@
+{% extends "public_profiles/base.html" %}
+{% load crispy_forms_tags %}
+{% block content %}
+<div class="profile-items-container">
+
+    <div class="profile-item">
+        <h1>{{ object.user }}</h1>
+        <p>Skills: {{ object.skills }}</p>
+        <p>Languages: {{ object.languages }}</p>
+        <p>{{ object.user }} doesn't like {{ object.dislikes}}</p>
+        <p>{{ object.user }} likes {{ object.likes }}</p>
+        <p>Github: {{ object.github }}</p>
+
+
+        <h2>Contact {{ object.user }}</h1>
+        <form method="post">
+            {% csrf_token %}
+            <fieldset>
+                {{ form|crispy }}
+                <h4>Where oh where is my form gone??</h4>
+            </fieldset>
+
+            <div class="form-actions">
+                <button type="submit">Send</button>
+            </div>
+        </form>
+
+    </div>
+
+</div>
+{% endblock content %}
