--- sendemail/templates/sendemail/email.html
@@ -1,13 +1,0 @@
-{% extends "public_profiles/base.html" %}
-{% block content %}
-
-    <h1>Contact Us</h1>
-    <form method="post">
-        {% csrf_token %}
-        {{ form.as_p }}
-        <div class="form-actions">
-        <button type="submit">Send</button>
-        </div>
-    </form>
-
-{% endblock content%}
